Add files via upload
Detecting and classifying software bugs are vital processes in the software development lifecycle, significantly affecting the quality and dependability of software systems. People can classify bugs, but that is slow and susceptible to error, especially as bugs become ever more complex in the modern age. Automated malfunction classification based on machine learning (ML) is a promising method to overcome those challenges.
Aiming to explore the development of an ML powered bug classification system.